+++ index.mjs
@@ -24,5 +24,13 @@ export const {
     softmax,
     linear,
     detectGPU,
-    templates
+    templates,
+    tokenize,
+    buildWord2Id,
+    buildVocab,
+    Encode,
+    stepDecay,
+    exponentialDecay,
+    cosineAnnealing,
+    reduceOnPlateau,
 } = pkg;