+++ layers/layer_functions/connectedLayer.js
@@ -141,16 +141,16 @@ const getOutputLayerDelta = (preds, actuals, zs, lossFunc, tasktype, layerObj) =
     }
     else if (tasktype === "multi_class_classification" && lossFunc === "sparse_categorical_cross_entropy") {
         dOutputLayer.set(preds);
+        if (!dOutputLayer[actuals[0]]) {
+            throw new Error(`Actual index value not exist in range. Actual target label: ${actuals[0]} | Output layer size: ${preds.length}`)
+        }
         dOutputLayer[actuals[0]] -= 1;
                         
     }
     else if (tasktype === "regression") {
         if (preds.length != actuals.length) throw new Error("Predictions array is not equal to actuals array");
 
-        const lastLayerZs = zs[zs.length - 1]; 
-        const dAct = dActivation(lastLayerZs); 
-
-        dOutputLayer = scaleDiff(preds, actuals, dAct);
+        dOutputLayer = element_wise_sub(preds, actuals);
 
         if (dOutputLayer.some(v => Number.isNaN(v))) throw new Error("Delta of the output layer has NaNs");